Reference

Reference

Modeling Languages

Addresses of Random Choices

Dynamic DSL

Static DSL

Combinators

Assignments

An assignment is a map from addresses of random choices to their values. Assignments are constructed by users to express observations and/or constraints on the traces of generative functions. Assignments are also returned by certain Gen inference methods, and are used internally by various Gen inference methods.

There are various concrete types for assignments, each of which is a subtype of Assignment. Assignments provide the following methods:

Gen.has_value — Function.

has_value(assmt::Assignment, addr)

Return true if there is a value at the given address.

source

Gen.get_value — Function.

value = get_value(assmt::Assignment, addr)

Return the value at the given address in the assignment, or throw a KeyError if no value exists. A syntactic sugar is Base.getindex:

value = assmt[addr]

source

Gen.get_subassmt — Function.

subassmt = get_subassmt(assmt::Assignment, addr)

Return the sub-assignment containing all choices whose address is prefixed by addr.

It is an error if the assignment contains a value at the given address. If there are no choices whose address is prefixed by addr then return an EmptyAssignment.

source

Gen.get_values_shallow — Function.

key_subassmt_iterable = get_values_shallow(assmt::Assignment)

Return an iterable collection of tuples (key, subassmt::Assignment) for each top-level key associated with a value.

source

Gen.get_subassmts_shallow — Function.

key_subassmt_iterable = get_subassmts_shallow(assmt::Assignment)

Return an iterable collection of tuples (key, subassmt::Assignment) for each top-level key that has a non-empty sub-assignment.

source

Gen.to_array — Function.

arr::Vector{T} = to_array(assmt::Assignment, ::Type{T}) where {T}

Populate an array with values of choices in the given assignment.

It is an error if each of the values cannot be coerced into a value of the given type.

Implementation

To support to_array, a concrete subtype T <: Assignment should implement the following method:

n::Int = _fill_array!(assmt::T, arr::Vector{V}, start_idx::Int) where {V}

Populate arr with values from the given assignment, starting at start_idx, and return the number of elements in arr that were populated.

source

Gen.from_array — Function.

assmt::Assignment = from_array(proto_assmt::Assignment, arr::Vector)

Return an assignment with the same address structure as a prototype assignment, but with values read off from the given array.

The order in which addresses are populated is determined by the prototype assignment. It is an error if the number of choices in the prototype assignment is not equal to the length the array.

Implementation

To support from_array, a concrete subtype T <: Assignment should implement the following method:

(n::Int, assmt::T) = _from_array(proto_assmt::T, arr::Vector{V}, start_idx::Int) where {V}

Return an assignment with the same address structure as a prototype assignment, but with values read off from arr, starting at position start_idx, and the number of elements read from arr.

source

Note that none of these methods mutate the assignment.

Assignments also provide Base.isempty, which tests of there are no random choices in the assignment, and Base.merge, which takes two assignments, and returns a new assignment containing all random choices in either assignment. It is an error if the assignments both have values at the same address, or if one assignment has a value at an address that is the prefix of the address of a value in the other assignment.

Dynamic Assignment

One concrete assignment type is DynamicAssignment, which is mutable. Users construct DynamicAssignments and populate them for use as observations or constraints, e.g.:

assmt = DynamicAssignment()
assmt[:x] = true
assmt["foo"] = 1.25
assmt[:y => 1 => :z] = -6.3

Gen.DynamicAssignment — Type.

assmt = DynamicAssignment()

Construct an empty dynamic assignment.

source

Gen.set_value! — Function.

set_value!(assmt::DynamicAssignment, addr, value)

Set the given value for the given address.

Will cause any previous value or sub-assignment at this address to be deleted. It is an error if there is already a value present at some prefix of the given address.

The following syntactic sugar is provided:

assmt[addr] = value

source

Gen.set_subassmt! — Function.

set_subassmt!(assmt::DynamicAssignment, addr, subassmt::Assignment)

Replace the sub-assignment rooted at the given address with the given sub-assignment. Set the given value for the given address.

Will cause any previous value or sub-assignment at the given address to be deleted. It is an error if there is already a value present at some prefix of address.

source

Selections

A selection is a set of addresses. Users typically construct selections and pass them to Gen inference library methods.

There are various concrete types for selections, each of which is a subtype of AddressSet. One such concrete type is DynamicAddressSet, which users can populate using Base.push!, e.g.:

sel = DynamicAddressSet()
push!(sel, :x)
push!(sel, "foo")
push!(sel, :y => 1 => :z)

There is also the following syntactic sugar:

sel = select(:x, "foo", :y => 1 => :z)

Inference Library

Importance Sampling

Gen.importance_sampling — Function.

(traces, log_norm_weights, lml_est) = importance_sampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment, num_samples::Int)

(traces, log_norm_weights, lml_est) = importance_sampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment,
    proposal::GenerativeFunction, proposal_args::Tuple,
    num_samples::Int)

Run importance sampling, returning a vector of traces with associated log weights.

The log-weights are normalized. Also return the estimate of the marginal likelihood of the observations (lml_est). The observations are addresses that must be sampled by the model in the given model arguments. The first variant uses the internal proposal distribution of the model. The second variant uses a custom proposal distribution defined by the given generative function. All addresses of random choices sampled by the proposal should also be sampled by the model function.

source

Gen.importance_resampling — Function.

(trace, lml_est) = importance_resampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment, num_samples::Int)

(traces, lml_est) = importance_resampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment,
    proposal::GenerativeFunction, proposal_args::Tuple,
    num_samples::Int)

Run sampling importance resampling, returning a single trace.

Unlike importance_sampling, the memory used constant in the number of samples.

source

Markov Chain Monte Carlo

Gen.default_mh — Function.

(new_trace, accepted) = default_mh(trace, selection::AddressSet)

Perform a Metropolis-Hastings update that proposes new values for the selected addresses from the internal proposal (often using ancestral sampling).

source

Gen.simple_mh — Function.

(new_trace, accepted) = simple_mh(trace, proposal::GenerativeFunction, proposal_args::Tuple)

Perform a Metropolis-Hastings update that proposes new values for some subset of random choices in the given trace using the given proposal generative function.

The proposal generative function should take as its first argument the current trace of the model, and remaining arguments proposal_args. All addresses sampled by the proposal must be in the existing model trace. The proposal may modify the control flow of the model, but values of new addresses in the model are sampled from the model's internal proposal distribution.

source

Gen.custom_mh — Function.

(new_trace, accepted) = custom_mh(trace, proposal::GenerativeFunction, proposal_args::Tuple)

Perform a Metropolis-Hastings update that proposes new values for some subset of random choices in the given trace using the given proposal generative function.

The proposal generative function should take as its first argument the current trace of the model, and remaining arguments proposal_args. If the proposal modifies addresses that determine the control flow in the model, values must be provided by the proposal for any addresses that are newly sampled by the model.

source

Gen.general_mh — Function.

(new_trace, accepted) = general_mh(trace, proposal, proposal_args, bijection)

Apply a MCMC update constructed from a bijection and a proposal that satisfies detailed balance.

source

Gen.mala — Function.

(new_trace, accepted) = mala(trace, selection::AddressSet, tau::Real)

Apply a Metropolis-Adjusted Langevin Algorithm (MALA) update.

Reference URL

source

Gen.hmc — Function.

(new_trace, accepted) = hmc(trace, selection::AddressSet, mass=0.1, L=10, eps=0.1)

Apply a Hamiltonian Monte Carlo (HMC) update.

Neal, Radford M. "MCMC using Hamiltonian dynamics." Handbook of Markov Chain Monte Carlo 2.11 (2011): 2.

Reference URL

source

Optimization over Random Choices

Gen.map_optimize — Function.

new_trace = map_optimize(trace, selection::AddressSet, 
    max_step_size=0.1, tau=0.5, min_step_size=1e-16, verbose=false)

Perform backtracking gradient ascent to optimize the log probability of the trace over selected continuous choices.

Selected random choices must have support on the entire real line.

source

Particle Filtering

Gen.particle_filter_default — Function.

(traces, log_norm_weights, lml_est) = particle_filter_default(
    model::GenerativeFunction, model_args::Tuple, num_steps::Int,
    num_particles::Int, ess_threshold::Real,
    init_observations::Assignment, step_observations::Function;
    verbose=false)

Run particle filtering using the internal proposal of the model.

The first argument to the model must be an integer, starting with 1, that defines the step. The remaining arguments are given by model_args. The model traces will be initialized with step=1 using the constraints given by init_observations. Then, the step will be consecutively incremented by 1. The function step_observations takes the step and returns a tuple (observations, argdiff) where observations is an assignment containing the values for newly observed random choices for the step, and argdiff describes the argument change from the previous step to the current step.

source

Gen.particle_filter_custom — Function.

(traces, log_norm_weights, lml_est) = particle_filter_custom(
    model::GenerativeFunction, model_args::Tuple, num_steps::Int,
    num_steps::Int, num_particles::Int, ess_threshold::Real,
    init_observations::Assignment, init_proposal_args::Tuple,
    step_observations::Function, step_proposal_args::Function,
    init_proposal::GenerativeFunction, step_proposal::GenerativeFunction;
    verbose::Bool=false)

Run particle filtering using custom proposal(s) at each step.

source

Training Generative Functions

Gen.sgd_train_batch — Function.

sgd_train_batch(teacher::GenerativeFunction, teacher_args::Tuple,
    batch_student::GenerativeFunction, conf::SGDTrainConf; verbose=false)

Train a batched (vectorized) student generative function on assignments sampled from a teacher generative function.

source

Generative Function Interface

A trace is a record of an execution of a generative function. There is no abstract type representing all traces. Generative functions implement the generative function interface, which is a set of methods that involve the execution traces and probabilistic behavior of generative functions. In the mathematical description of the interface methods, we denote arguments to a function by $x$, complete assignments of values to addresses of random choices (containing all the random choices made during some execution) by $t$ and partial assignments by either $u$ or $v$. We denote a trace of a generative function by the tuple $(x, t)$. We say that two assignments $u$ and $t$ agree when they assign addresses that appear in both assignments to the same values (they can different or even disjoint sets of addresses and still agree). A generative function is associated with a family of probability distributions $P(t; x)$ on assignments $t$, parameterized by arguments $x$, and a second family of distributions $Q(t; u, x)$ on assignments $t$ parameterized by partial assignment $u$ and arguments $x$. $Q$ is called the internal proposal family of the generative function, and satisfies that if $u$ and $t$ agree then $P(t; x) > 0$ if and only if $Q(t; x, u) > 0$, and that $Q(t; x, u) > 0$ implies that $u$ and $t$ agree. See the Gen technical report for additional details.

Generative functions may also use non-addressable random choices, denoted $r$. Unlike regular (addressable) random choices, non-addressable random choices do not have addresses, and the value of non-addressable random choices is not exposed through the generative function interface. However, the state of non-addressable random choices is maintained in the trace. A trace that contains non-addressable random choices is denoted $(x, t, r)$. Non-addressable random choices manifest to the user of the interface as stochasticity in weights returned by generative function interface methods. The behavior of non-addressable random choices is defined by an additional pair of families of distributions associated with the generative function, denoted $Q(r; x, t)$ and $P(r; x, t)$, which are defined for $P(t; x) > 0$, and which satisfy $Q(r; x, t) > 0$ if and only if $P(r; x, t) > 0$. For each generative function below, we describe its semantics first in the basic setting where there is no non-addressable random choices, and then in the more general setting that may include non-addressable random choices.

Gen.initialize — Function.

(trace::U, weight) = initialize(gen_fn::GenerativeFunction{T,U}, args::Tuple,
                                assmt::Assignment)

Return a trace of a generative function that is consistent with the given assignment.

Basic case

Given arguments $x$ (args) and assignment $u$ (assmt), sample $t \sim Q(\cdot; u, x)$ and return the trace $(x, t)$ (trace). Also return the weight (weight):

\[\frac{P(t; x)}{Q(t; u, x)}\]

General case

Identical to the basic case, except that we also sample $r \sim Q(\cdot; x, t)$, the trace is $(x, t, r)$ and the weight is:

\[\frac{P(t; x)}{Q(t; u, x)} \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.project — Function.

weight = project(trace::U, selection::AddressSet)

Estimate the probability that the selected choices take the values they do in a trace.

Basic case

Given a trace $(x, t)$ (trace) and a set of addresses $A$ (selection), let $u$ denote the restriction of $t$ to $A$. Return the weight (weight):

\[\frac{P(t; x)}{Q(t; u, x)}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$ and the weight is:

\[\frac{P(t; x)}{Q(t; u, x)} \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.propose — Function.

(assmt, weight, retval) = propose(gen_fn::GenerativeFunction, args::Tuple)

Sample an assignment and compute the probability of proposing that assignment.

Basic case

Given arguments (args), sample $t' \sim P(\cdot; x)$, and return $t$ (assmt) and the weight (weight) $P(t; x)$.

General case

Identical to the basic case, except that we also sample $r \sim P(\cdot; x, t)$, and the weight is:

\[P(t; x) \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.assess — Function.

(weight, retval) = assess(gen_fn::GenerativeFunction, args::Tuple, assmt::Assignment)

Return the probability of proposing an assignment

Basic case

Given arguments $x$ (args) and an assignment $t$ (assmt) such that $P(t; x) > 0$, return the weight (weight) $P(t; x)$. It is an error if $P(t; x) = 0$.

General case

Identical to the basic case except that we also sample $r \sim Q(\cdot; x, t)$, and the weight is:

\[P(t; x) \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.force_update — Function.

(new_trace, weight, discard, retdiff) = force_update(args::Tuple, argdiff, trace,
                                                     assmt::Assignment)

Update a trace by changing the arguments and/or providing new values for some existing random choice(s) and values for any newly introduced random choice(s).

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and an assignment $u$ (assmt), return a new trace $(x', t')$ (new_trace) that is consistent with $u$. The values of choices in $t'$ are deterministically copied either from $t$ or from $u$ (with $u$ taking precedence). All choices in $u$ must appear in $t'$. Also return an assignment $v$ (discard) containing the choices in $t$ that were overwritten by values from $u$, and any choices in $t$ whose address does not appear in $t'$. Also return the weight (weight):

\[\frac{P(t'; x')}{P(t; x)}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, the new trace is $(x', t', r')$ where $r' \sim Q(\cdot; x', t')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.fix_update — Function.

(new_trace, weight, discard, retdiff) = fix_update(args::Tuple, argdiff, trace,
                                                   assmt::Assignment)

Update a trace, by changing the arguments and/or providing new values for some existing random choice(s).

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and an assignment $u$ (assmt), return a new trace $(x', t')$ (new_trace) that is consistent with $u$. Let $u + t$ denote the merge of $u$ and $t$ (with $u$ taking precedence). Sample $t' \sim Q(\cdot; u + t, x)$. All addresses in $u$ must appear in $t$ and in $t'$. Also return an assignment $v$ (discard) containing the values from $t$ for addresses in $u$. Also return the weight (weight):

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; v + t', x)}{Q(t'; u + t, x')}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, the new trace is $(x', t', r')$ where $r' \sim Q(\cdot; x', t')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; v + t', x)}{Q(t'; u + t, x')} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.free_update — Function.

(new_trace, weight, retdiff) = free_update(args::Tuple, argdiff, trace,
                                           selection::AddressSet)

Update a trace by changing the arguments and/or randomly sampling new values for selected random choices.

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and a set of addresses $A$ (selection), return a new trace $(x', t')$ (new_trace) such that $t'$ agrees with $t$ on all addresses not in $A$ ($t$ and $t'$ may have different sets of addresses). Let $u$ denote the restriction of $t$ to the complement of $A$. Sample $t' \sim Q(\cdot; u, x')$. Return the new trace $(x', t')$ (new_trace) and the weight (weight):

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; u', x)}{Q(t'; u, x')}\]

where $u'$ is the restriction of $t'$ to the complement of $A$.

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, the new trace is $(x', t', r')$ where $r' \sim Q(\cdot; x', t')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; u', x)}{Q(t'; u, x')} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.extend — Function.

(new_trace, weight, retdiff) = extend(args::Tuple, argdiff, trace, assmt::Assignment)

Extend a trace with new random choices by changing the arguments.

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and an assignment $u$ (assmt) that shares no addresses with $t$, return a new trace $(x', t')$ (new_trace) such that $t'$ agrees with $t$ on all addresses in $t$ and $t'$ agrees with $u$ on all addresses in $u$. Sample $t' \sim Q(\cdot; t + u, x')$. Also return the weight (weight):

\[\frac{P(t'; x')}{P(t; x) Q(t'; t + u, x')}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, and we also sample $r' \sim Q(\cdot; t', x)$, the new trace is $(x', t', r')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x) Q(t'; t + u, x')} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.backprop_params — Function.

arg_grads = backprop_params(trace, retgrad)

Increment gradient accumulators for parameters by the gradient of the log-probability of the trace.

Basic case

Given a previous trace $(x, t)$ (trace) and a gradient with respect to the return value $∇_y J$ (retgrad), return the following gradient (arg_grads) with respect to the arguments $x$:

\[∇_x \left( \log P(t; x) + J \right)\]

Also increment the gradient accumulators for the static parameters $Θ$ of the function by:

\[∇_Θ \left( \log P(t; x) + J \right)\]

General case

Not yet formalized.

source

Gen.backprop_trace — Function.

(arg_grads, choice_values, choice_grads) = backprop_trace(trace, selection::AddressSet,
                                                          retgrad)

Basic case

Given a previous trace $(x, t)$ (trace) and a gradient with respect to the return value $∇_y J$ (retgrad), return the following gradient (arg_grads) with respect to the arguments $x$:

\[∇_x \left( \log P(t; x) + J \right)\]

Also given a set of addresses $A$ (selection) that are continuous-valued random choices, return the folowing gradient (choice_grads) with respect to the values of these choices:

\[∇_A \left( \log P(t; x) + J \right)\]

Also return the assignment (choice_values) that is the restriction of $t$ to $A$.

General case

Not yet formalized.

source

Gen.get_assmt — Function.

get_assmt(trace)

Return a value implementing the assignment interface

source

Gen.get_args — Function.

get_args(trace)

Return the argument tuple for a given execution.

source

Gen.get_retval — Function.

get_retval(trace)

Return the return value of the given execution.

source

Gen.get_score — Function.

get_score(trace)

Basic case

Return $P(t; x)$

General case

Return $P(r, t; x) / Q(r; tx, t)$

source