Probability Distributions

Probability Distributions

Built-In Distributions

Gen.bernoulli — Constant.

bernoulli(prob_true::Real)

Samples a Bool value which is true with given probability

source

Gen.normal — Constant.

normal(mu::Real, std::Real)

Samples a Float64 value from a normal distribution.

source

Gen.mvnormal — Constant.

mvnormal(mu::AbstractVector{T}, cov::AbstractMatrix{U}} where {T<:Real,U<:Real}

Samples a Vector{Float64} value from a multivariate normal distribution.

source

Gen.gamma — Constant.

gamma(shape::Real, scale::Real)

Sample a Float64 from a gamma distribution.

source

Gen.inv_gamma — Constant.

inv_gamma(shape::Real, scale::Real)

Sample a Float64 from a inverse gamma distribution.

source

Gen.beta — Constant.

beta(alpha::Real, beta::Real)

Sample a Float64 from a beta distribution.

source

Gen.categorical — Constant.

categorical(probs::AbstractArray{U, 1}) where {U <: Real}

Given a vector of probabilities probs where sum(probs) = 1, sample an Int i from the set {1, 2, .., length(probs)} with probability probs[i].

source

Gen.uniform — Constant.

uniform(low::Real, high::Real)

Sample a Float64 from the uniform distribution on the interval [low, high].

source

Gen.uniform_discrete — Constant.

uniform_discrete(low::Integer, high::Integer)

Sample an Int from the uniform distribution on the set {low, low + 1, ..., high-1, high}.

source

Gen.poisson — Constant.

poisson(lambda::Real)

Sample an Int from the Poisson distribution with rate lambda.

source

Defining New Distributions

Probability distributions are singleton types whose supertype is Distribution{T}, where T indicates the data type of the random sample.

abstract type Distribution{T} end

By convention, distributions have a global constant lower-case name for the singleton value. For example:

struct Bernoulli <: Distribution{Bool} end
const bernoulli = Bernoulli()

Distributions must implement two methods, random and logpdf.

random returns a random sample from the distribution:

x::Bool = random(bernoulli, 0.5)
x::Bool = random(Bernoulli(), 0.5)

logpdf returns the log probability (density) of the distribution at a given value:

logpdf(bernoulli, false, 0.5)
logpdf(Bernoulli(), false, 0.5)

Distribution values are also callable, which is a syntactic sugar with the same behavior of calling random:

bernoulli(0.5) # identical to random(bernoulli, 0.5) and random(Bernoulli(), 0.5)