Generative Functions

Generative Functions

One of the core abstractions in Gen is the generative function. Generative functions are used to represent a variety of different types of probabilistic computations including generative models, inference models, custom proposal distributions, and variational approximations.

Generative functions are represented by the following abstact type:

Gen.GenerativeFunction — Type.

GenerativeFunction{T,U}

Abstract type for a generative function with return value type T and trace type U.

source

There are various kinds of generative functions, which are represented by concrete subtypes of GenerativeFunction. For example, the Built-in Modeling Language allows generative functions to be constructed using Julia function definition syntax:

@gen function foo(a, b)
    if @addr(bernoulli(0.5), :z)
        return a + b + 1
    else
        return a + b
    end
end

Generative functions must implement the the methods of the Generative Function Interface. Generative functions behave like Julia functions in some respects. For example, we can call a generative function foo on arguments and get an output value using regular Julia call syntax:

>julia foo(2, 4)
7

However, generative functions are distinct from Julia functions because they support additional behaviors, described in the remainder of this section.

Probabilistic semantics

Generative functions may use randomness, but they may not mutate externally observable state. We represent the randomness used during an execution as a map from unique addresses to values, denoted $t : A \to V$ where $A$ is an address set and $V$ is a set of possible values that random choices can take. In this section, we assume that random choices are discrete to simplify notation. We say that two random choice maps $t$ and $s$ agree if they assign the same value for all addresses that is in both of their domains. Formally, every generative function is associated with three mathematical objects:

Input type

The set of valid argument tuples to the function, denoted $X$.

Probability distribution family

A family of probability distributions $p(t; x)$ on maps $t$ from random choice addresses to their values, indexed by arguments $x$, for all $x \in X$. Note that the distribution must be normalized:

\[\sum_{t} p(t; x) = 1 \;\; \mbox{for all} \;\; x \in X\]

This corresponds to a requirement that the function terminate with probabability 1 for all valid arguments.

Return value function

A (deterministic) function $f$ that maps the tuple $(x, t)$ of the arguments and the random choice map to the return value of the function (which we denote by $y$).

Internal proposal distribution family

A family of probability distributions $q(t; x, u)$ on maps $t$ from random choice addresses to their values, indexed by tuples $(x, u)$ where $u$ is a map from random choice addresses to values, and where $x$ are the arguments to the function. It must be normalized for all valid $(x, u)$:

\[\sum_{t} q(t; x, u) = 1 \;\; \mbox{for all} \;\; x \in X, u\]

Also, it must satisfy the following condition:

\[p(t; x) > 0 \mbox{ if and only if } q(t; x, u) > 0 \mbox{ for all } u \mbox{ where } u \mbox{ and } t \mbox{ agree }\]

Finally, we require that:

\[q(t; x, u) > 0 \mbox{ implies that } u \mbox{ and } t \mbox{ agree }.\]

Execution traces

An execution trace (or just trace) is a record of an execution of a generative function. There is no abstract type representing all traces. Different concrete types of generative functions use different data structures and different Jula types for their traces. The trace type that a generative function uses is the second type parameter of the GenerativeFunction abstract type.

A trace of a generative function can be produced using initialize:

(trace, weight) = initialize(foo, (2, 4))

The trace contains various information about the execution, including:

The arguments to the function, which is retrieved with get_args:

>julia get_args(trace)
(2, 4)

The return value of the function, which is retrieved with get_retval:

>julia get_retval(trace)
7

The map from addresses of random choices to their values, which is retrieved with get_assmt, and has abstract type Assignment.

>julia println(get_assmt(trace))
│
└── :z : false

The log probability that the random choices took the values they did for the arguments, available with get_score:

>julia get_score(trace)
-0.6931471805599453

A reference to the generative function that was executed, which is retrieved with get_gen_fn:

>julia foo === get_gen_fn(trace)
true

Trace update methods

There are several methods that take a trace of a generative function as input and return a new trace of the generative function based on adjustments to the execution history of the function:

force_update

fix_update

free_update

extend

In addition to the input trace, and other arguments that indicate how to adjust the trace, each of these methods also accepts an args argument and an argdiff argument. The args argument contains the new arguments to the generative function, which may differ from the previous arguments to the generative function (which can be retrieved by applying get_args to the previous trace). In many cases, the adjustment to the execution specified by the other arguments to these methods is 'small' and only effects certain parts of the computation. Therefore, it is often possible to generate the new trace and the appropriate log probability ratios required for these methods without revisiting every state of the computation of the generative function. To enable this, the argdiff argument provides additional information about the difference between the previous arguments to the generative function, and the new arguments. This argdiff information permits the implementation of the update method to avoid inspecting the entire argument data structure to identify which parts were updated.

Differentiable programming

Generative functions may support computation of gradients with respect to (i) all or a subset of its arguments, (ii) its trainable parameters, and (iii) the value of certain random choices. The set of elements (either arguments, trainable parameters, or random choices) for which gradients are available is called the gradient source set. A generative function statically reports whether or not it is able to compute gradients with respect to each of its arguments, through the function has_argument_grads. Let $x_G$ denote the set of arguments for which the generative function does support gradient computation. Similarly, a generative function supports gradients with respect the value of random choices made at all or a subset of addresses. If the return value of the function is conditionally independent of each element in the gradient source set given the other elements in the gradient source set and values of all other random choices, for all possible traces of the function, then the generative function requires a return value gradient to compute gradients with respect to elements of the gradient source set. This static property of the generative function is reported by accepts_output_grad.

Trainable parameters

Non-addressed randomness

Generative functions may also use non-addressable randomness, which is not returned by get_assmt. However, the state of non-addressable random choices is maintained by the trace internally. We denote non-addressable randomness by $r$. The probabilistic semantics are extended for a generative function with non-addressable randomness as follows:

Input type

No extension necessary

Probability distribution family

A family of probability distributions $p(t, r; x)$ that is normalized for all $x \in X$, that factors according to:

\[p(t, r; x) = p(t; x) p(r; t, x)\]

Return value function

The return value cannot be a function of the non-addressable randomness. It remains a function $f$ on tuples $(x, t)$.

Internal proposal distribution family

A family of distributions $q(t, r; x, u)$ that factors according to:

\[q(t, r; x, u) = q(t; x, u) q(r; x, t)\]

where $q(t; x, u)$ satisfies the conditions stated above. Note that the distribution on internal randomness does not depend on $u$ in this factorization.

Generative Function Interface

Gen.has_argument_grads — Function.

bools::Tuple = has_argument_grads(gen_fn::Union{GenerativeFunction,Distribution})

Return a tuple of booleans indicating whether a gradient is available for each of its arguments.

source

Gen.accepts_output_grad — Function.

req::Bool = accepts_output_grad(gen_fn::GenerativeFunction)

Return a boolean indicating whether the return value is dependent on any of the gradient source elements for any trace.

The gradient source elements are:

Any argument whose position is true in has_argument_grads

Any static parameter

Random choices made at a set of addresses that are selectable by backprop_trace.

source

Gen.initialize — Function.

(trace::U, weight) = initialize(gen_fn::GenerativeFunction{T,U}, args::Tuple,
                                assmt::Assignment)

Return a trace of a generative function that is consistent with the given assignment.

Basic case

Given arguments $x$ (args) and assignment $u$ (assmt), sample $t \sim Q(\cdot; u, x)$ and return the trace $(x, t)$ (trace). Also return the weight (weight):

\[\frac{P(t; x)}{Q(t; u, x)}\]

General case

Identical to the basic case, except that we also sample $r \sim Q(\cdot; x, t)$, the trace is $(x, t, r)$ and the weight is:

\[\frac{P(t; x)}{Q(t; u, x)} \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.project — Function.

weight = project(trace::U, selection::AddressSet)

Estimate the probability that the selected choices take the values they do in a trace.

Basic case

Given a trace $(x, t)$ (trace) and a set of addresses $A$ (selection), let $u$ denote the restriction of $t$ to $A$. Return the weight (weight):

\[\frac{P(t; x)}{Q(t; u, x)}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$ and the weight is:

\[\frac{P(t; x)}{Q(t; u, x)} \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.propose — Function.

(assmt, weight, retval) = propose(gen_fn::GenerativeFunction, args::Tuple)

Sample an assignment and compute the probability of proposing that assignment.

Basic case

Given arguments (args), sample $t' \sim P(\cdot; x)$, and return $t$ (assmt) and the weight (weight) $P(t; x)$.

General case

Identical to the basic case, except that we also sample $r \sim P(\cdot; x, t)$, and the weight is:

\[P(t; x) \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.assess — Function.

(weight, retval) = assess(gen_fn::GenerativeFunction, args::Tuple, assmt::Assignment)

Return the probability of proposing an assignment

Basic case

Given arguments $x$ (args) and an assignment $t$ (assmt) such that $P(t; x) > 0$, return the weight (weight) $P(t; x)$. It is an error if $P(t; x) = 0$.

General case

Identical to the basic case except that we also sample $r \sim Q(\cdot; x, t)$, and the weight is:

\[P(t; x) \cdot \frac{P(r; x, t)}{Q(r; x, t)}\]

source

Gen.force_update — Function.

(new_trace, weight, discard, retdiff) = force_update(args::Tuple, argdiff, trace,
                                                     assmt::Assignment)

Update a trace by changing the arguments and/or providing new values for some existing random choice(s) and values for any newly introduced random choice(s).

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and an assignment $u$ (assmt), return a new trace $(x', t')$ (new_trace) that is consistent with $u$. The values of choices in $t'$ are deterministically copied either from $t$ or from $u$ (with $u$ taking precedence). All choices in $u$ must appear in $t'$. Also return an assignment $v$ (discard) containing the choices in $t$ that were overwritten by values from $u$, and any choices in $t$ whose address does not appear in $t'$. Also return the weight (weight):

\[\frac{P(t'; x')}{P(t; x)}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, the new trace is $(x', t', r')$ where $r' \sim Q(\cdot; x', t')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.fix_update — Function.

(new_trace, weight, discard, retdiff) = fix_update(args::Tuple, argdiff, trace,
                                                   assmt::Assignment)

Update a trace, by changing the arguments and/or providing new values for some existing random choice(s).

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and an assignment $u$ (assmt), return a new trace $(x', t')$ (new_trace) that is consistent with $u$. Let $u + t$ denote the merge of $u$ and $t$ (with $u$ taking precedence). Sample $t' \sim Q(\cdot; u + t, x)$. All addresses in $u$ must appear in $t$ and in $t'$. Also return an assignment $v$ (discard) containing the values from $t$ for addresses in $u$. Also return the weight (weight):

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; v + t', x)}{Q(t'; u + t, x')}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, the new trace is $(x', t', r')$ where $r' \sim Q(\cdot; x', t')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; v + t', x)}{Q(t'; u + t, x')} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.free_update — Function.

(new_trace, weight, retdiff) = free_update(args::Tuple, argdiff, trace,
                                           selection::AddressSet)

Update a trace by changing the arguments and/or randomly sampling new values for selected random choices.

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and a set of addresses $A$ (selection), return a new trace $(x', t')$ (new_trace) such that $t'$ agrees with $t$ on all addresses not in $A$ ($t$ and $t'$ may have different sets of addresses). Let $u$ denote the restriction of $t$ to the complement of $A$. Sample $t' \sim Q(\cdot; u, x')$. Return the new trace $(x', t')$ (new_trace) and the weight (weight):

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; u', x)}{Q(t'; u, x')}\]

where $u'$ is the restriction of $t'$ to the complement of $A$.

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, the new trace is $(x', t', r')$ where $r' \sim Q(\cdot; x', t')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x)} \cdot \frac{Q(t; u', x)}{Q(t'; u, x')} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.extend — Function.

(new_trace, weight, retdiff) = extend(args::Tuple, argdiff, trace, assmt::Assignment)

Extend a trace with new random choices by changing the arguments.

Basic case

Given a previous trace $(x, t)$ (trace), new arguments $x'$ (args), and an assignment $u$ (assmt) that shares no addresses with $t$, return a new trace $(x', t')$ (new_trace) such that $t'$ agrees with $t$ on all addresses in $t$ and $t'$ agrees with $u$ on all addresses in $u$. Sample $t' \sim Q(\cdot; t + u, x')$. Also return the weight (weight):

\[\frac{P(t'; x')}{P(t; x) Q(t'; t + u, x')}\]

General case

Identical to the basic case except that the previous trace is $(x, t, r)$, and we also sample $r' \sim Q(\cdot; t', x)$, the new trace is $(x', t', r')$, and the weight is:

\[\frac{P(t'; x')}{P(t; x) Q(t'; t + u, x')} \cdot \frac{P(r'; x', t') Q(r; x, t)}{P(r; x, t) Q(r'; x', t')}\]

source

Gen.backprop_params — Function.

arg_grads = backprop_params(trace, retgrad, scaler=1.)

Increment gradient accumulators for parameters by the gradient of the log-probability of the trace, optionally scaled, and return the gradient with respect to the arguments (not scaled).

Basic case

Given a previous trace $(x, t)$ (trace) and a gradient with respect to the return value $∇_y J$ (retgrad), return the following gradient (arg_grads) with respect to the arguments $x$:

\[∇_x \left( \log P(t; x) + J \right)\]

Also increment the gradient accumulators for the static parameters $Θ$ of the function by:

\[∇_Θ \left( \log P(t; x) + J \right)\]

General case

Not yet formalized.

source

Gen.backprop_trace — Function.

(arg_grads, choice_values, choice_grads) = backprop_trace(trace, selection::AddressSet,
                                                          retgrad)

Basic case

Given a previous trace $(x, t)$ (trace) and a gradient with respect to the return value $∇_y J$ (retgrad), return the following gradient (arg_grads) with respect to the arguments $x$:

\[∇_x \left( \log P(t; x) + J \right)\]

Also given a set of addresses $A$ (selection) that are continuous-valued random choices, return the folowing gradient (choice_grads) with respect to the values of these choices:

\[∇_A \left( \log P(t; x) + J \right)\]

Also return the assignment (choice_values) that is the restriction of $t$ to $A$.

General case

Not yet formalized.

source

Gen.get_assmt — Function.

get_assmt(trace)

Return a value implementing the assignment interface

source

Gen.get_args — Function.

get_args(trace)

Return the argument tuple for a given execution.

source

Gen.get_retval — Function.

get_retval(trace)

Return the return value of the given execution.

source

Gen.get_score — Function.

get_score(trace)

Basic case

Return $P(t; x)$

General case

Return $P(r, t; x) / Q(r; tx, t)$

source

Gen.get_gen_fn — Function.

gen_fn::GenerativeFunction = get_gen_fn(trace)

Return the generative function that produced the given trace.

source

Gen.get_params — Function.

get_params(gen_fn::GenerativeFunction)

Return an iterable over the trainable parameters of the generative function.

source