Optimizing Static Parameters

Optimizing Static Parameters

Gradient-based optimization of the static parameters of generative functions is based on interleaving (1) accumulation of gradients with respect to static parameters using backprop_params with (2) updates to the static parameters that read from the gradients, mutate the value of the static parameters, and reset the gradients to zero.

Gen has built-in support for the following configuration types of updates:

Gen.GradientDescent — Type.

conf = GradientDescent(step_size_init, step_size_beta)

Configuration for stochastic gradient descent update with step size given by (t::Int) -> step_size_init * (step_size_beta + 1) / (step_size_beta + t) where t is the iteration number.

source

Gen.ADAM — Type.

conf = ADAM(learning_rate, beta1, beta2)

Configuration for ADAM update.

source

Parameter updates are constructed by combining a configuration value with the set of static parameters that it should be applied to:

Gen.ParamUpdate — Type.

update = ParamUpdate(conf, param_lists...)

Return an update configured by conf that applies to set of parameters defined by param_lists.

Each element in param_lists value is is pair of a generative function and a vector of its parameter references.

Example. To construct an update that applies a gradient descent update to the parameters :a and :b of generative function foo and the parameter :theta of generative function :bar:

update = ParamUpdate(GradientDescent(0.001, 100), foo => [:a, :b], bar => [:theta])

update = ParamUpdate(conf, gen_fn::GenerativeFunction)

Return an update configured by conf that applies to all trainable parameters owned by the given generative function.

Note that trainable parameters not owned by the given generative function will not be updated, even if they are used during execution of the function. This is syntactic sugar for the constructor form above.

Example. If generative function foo has parameters :a and :b, to construct an update that applies a gradient descent update to the parameters :a and :b:

update = ParamUpdate(GradientDescent(0.001, 100), foo)

source

Finally, an update is applied with:

Gen.apply! — Function.

apply!(update::ParamUpdate)

Perform one step of the update.

source