Inference Library

Inference Library

Importance Sampling

Gen.importance_sampling — Function.

(traces, log_norm_weights, lml_est) = importance_sampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment, num_samples::Int)

(traces, log_norm_weights, lml_est) = importance_sampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment,
    proposal::GenerativeFunction, proposal_args::Tuple,
    num_samples::Int)

Run importance sampling, returning a vector of traces with associated log weights.

The log-weights are normalized. Also return the estimate of the marginal likelihood of the observations (lml_est). The observations are addresses that must be sampled by the model in the given model arguments. The first variant uses the internal proposal distribution of the model. The second variant uses a custom proposal distribution defined by the given generative function. All addresses of random choices sampled by the proposal should also be sampled by the model function.

source

Gen.importance_resampling — Function.

(trace, lml_est) = importance_resampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment, num_samples::Int)

(traces, lml_est) = importance_resampling(model::GenerativeFunction,
    model_args::Tuple, observations::Assignment,
    proposal::GenerativeFunction, proposal_args::Tuple,
    num_samples::Int)

Run sampling importance resampling, returning a single trace.

Unlike importance_sampling, the memory used constant in the number of samples.

source

Markov Chain Monte Carlo

The following inference library methods take a trace and return a new trace.

Gen.default_mh — Function.

(new_trace, accepted) = default_mh(trace, selection::AddressSet)

Perform a Metropolis-Hastings update that proposes new values for the selected addresses from the internal proposal (often using ancestral sampling).

source

Gen.simple_mh — Function.

(new_trace, accepted) = simple_mh(trace, proposal::GenerativeFunction, proposal_args::Tuple)

Perform a Metropolis-Hastings update that proposes new values for some subset of random choices in the given trace using the given proposal generative function.

The proposal generative function should take as its first argument the current trace of the model, and remaining arguments proposal_args. All addresses sampled by the proposal must be in the existing model trace. The proposal may modify the control flow of the model, but values of new addresses in the model are sampled from the model's internal proposal distribution.

source

Gen.custom_mh — Function.

(new_trace, accepted) = custom_mh(trace, proposal::GenerativeFunction, proposal_args::Tuple)

Perform a Metropolis-Hastings update that proposes new values for some subset of random choices in the given trace using the given proposal generative function.

The proposal generative function should take as its first argument the current trace of the model, and remaining arguments proposal_args. If the proposal modifies addresses that determine the control flow in the model, values must be provided by the proposal for any addresses that are newly sampled by the model.

source

Gen.general_mh — Function.

(new_trace, accepted) = general_mh(trace, proposal::GenerativeFunction, proposal_args::Tuple, involution::Function)

Perform a generalized Metropolis-Hastings update based on an involution (bijection that is its own inverse) on a space of assignments.

The `involution' Julia function has the following signature:

(new_trace, bwd_assmt::Assignment, weight) = involution(trace, fwd_assmt::Assignment, fwd_ret, proposal_args::Tuple)

The generative function proposal is executed on arguments (trace, proposal_args...), producing an assignment fwd_assmt and return value fwd_ret. For each value of model arguments (contained in trace) and proposal_args, the involution function applies an involution that maps the tuple (get_assmt(trace), fwd_assmt) to the tuple (get_assmt(new_trace), bwd_assmt). Note that fwd_ret is a deterministic function of fwd_assmt and proposal_args. When only discrete random choices are used, the weight must be equal to get_score(new_trace) - get_score(trace).

Including Continuous Random Choices When continuous random choices are used, the weight must include an additive term that is the determinant of the the Jacobian of the bijection on the continuous random choices that is obtained by currying the involution on the discrete random choices.

source

Gen.mala — Function.

(new_trace, accepted) = mala(trace, selection::AddressSet, tau::Real)

Apply a Metropolis-Adjusted Langevin Algorithm (MALA) update.

Reference URL

source

Gen.hmc — Function.

(new_trace, accepted) = hmc(trace, selection::AddressSet, mass=0.1, L=10, eps=0.1)

Apply a Hamiltonian Monte Carlo (HMC) update.

Neal, Radford M. "MCMC using Hamiltonian dynamics." Handbook of Markov Chain Monte Carlo 2.11 (2011): 2.

Reference URL

source

Optimization over Random Choices

Gen.map_optimize — Function.

new_trace = map_optimize(trace, selection::AddressSet, 
    max_step_size=0.1, tau=0.5, min_step_size=1e-16, verbose=false)

Perform backtracking gradient ascent to optimize the log probability of the trace over selected continuous choices.

Selected random choices must have support on the entire real line.

source

Particle Filtering

Gen.particle_filter_default — Function.

(traces, log_norm_weights, lml_est) = particle_filter_default(
    model::GenerativeFunction, model_args::Tuple, num_steps::Int,
    num_particles::Int, ess_threshold::Real,
    init_observations::Assignment, step_observations::Function;
    verbose=false)

Run particle filtering using the internal proposal of the model.

The first argument to the model must be an integer, starting with 1, that defines the step. The remaining arguments are given by model_args. The model traces will be initialized with step=1 using the constraints given by init_observations. Then, the step will be consecutively incremented by 1. The function step_observations takes the step and returns a tuple (observations, argdiff) where observations is an assignment containing the values for newly observed random choices for the step, and argdiff describes the argument change from the previous step to the current step.

source

Gen.particle_filter_custom — Function.

(traces, log_norm_weights, lml_est) = particle_filter_custom(
    model::GenerativeFunction, model_args::Tuple, num_steps::Int,
    num_steps::Int, num_particles::Int, ess_threshold::Real,
    init_observations::Assignment, init_proposal_args::Tuple,
    step_observations::Function, step_proposal_args::Function,
    init_proposal::GenerativeFunction, step_proposal::GenerativeFunction;
    verbose::Bool=false)

Run particle filtering using custom proposal(s) at each step.

source

Training Generative Functions

sgd_train_batch